Store API server temporary files in /var/tmp instead of /tmp.
We don't actually want files preserved across reboots, but on some systems files in /tmp are automatically purged (tmpwatch/systemd-tmpfiles), which would put the API server in an inconsistent state. Since we should clean up our files when shutting down only a crash or hard shutdown of the system should be able to leave garbage behind. 

Also remove a duplicate utility function in the process.